services/bin: check in prebuilt binaries
These binaries aren't very likely to change in the near future. While
checking in binaries into the repo is not a great thing to do, the
biggest issue with that is that nobody knows how to reproduce them.

Since we're using a (pretty well) reproducible recipe for generating
this (tools/build-binaries.sh run in a Docker container), this should
outweigh the cost of preparing the NDK and rebuilding them on every CI
build...